streamlit/requirements.txt
numpy
streamlit
tensorflow==2.6
pillow
matplotlib==3.3.4